requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
psycopg2-binary
python-multipart
python-dotenv
pydantic[email]
alembic